prep for parsing
