+++ back-end/loader/requirements.txt
@@ -1,2 +1,2 @@
 fitdecode
-pandas
+pymongo
