--- back-end/loader/schema.json
@@ -1,11 +1,0 @@
-{
-    _id: ObjectId("5effaa5662679b5af2c58829"),
-    username: "username"
-    activityId: 123455
-    activityDate: 2022-01-01
-    duration: 1234
-    distance: 1234
-    laps: [{index: 1, duration, lap_trigger}]
-    route: [{index: 1, timestamp: ts, lat:123, lon: 123}, {index: 2, timestamp: ts, lat:123, lon: 123}]
-              
-}
